Added roundtrip example
