+++ identity_jose/Cargo.toml
@@ -33,3 +33,8 @@ iota-crypto = { version = "0.15.3", features = ["ed25519", "random", "hmac"] }
 p256 = { version = "0.12.0", default-features = false, features = ["std", "ecdsa", "ecdsa-core"] }
 signature = { version = "2", default-features = false }
 tokio = { version = "1.24.2", features = ["macros", "rt"] }
+
+[[example]]
+name = "jws_encoding_decoding"
+publish = false
+test = true
